refactor: less detail namespace
