+++ include/blackhole/factory.hpp
@@ -22,5 +22,5 @@ public:
     virtual auto from(const config::node_t& config) const -> std::unique_ptr<T> = 0;
 };
 
-}  // namespace v1
-}  // namespace blackhole
+} // namespace v1
+} // namespace blackhole
