+++ src/formatter/json.cpp
@@ -21,10 +21,10 @@
 #include "blackhole/record.hpp"
 #include "blackhole/extensions/writer.hpp"
 
-#include "blackhole/detail/attribute.hpp"
 #include "blackhole/detail/memory.hpp"
 #include "blackhole/detail/util/deleter.hpp"
 
+#include "../attribute.hpp"
 #include "../datetime.hpp"
 #include "json.hpp"
