+++ src/formatter/tskv.cpp
@@ -12,10 +12,10 @@
 #include "blackhole/formatter.hpp"
 #include "blackhole/record.hpp"
 
-#include "blackhole/detail/attribute.hpp"
 #include "blackhole/detail/memory.hpp"
 #include "blackhole/detail/util/deleter.hpp"
 
+#include "../attribute.hpp"
 #include "../datetime.hpp"
 #include "string/token.hpp"
